Moved Directory to Example
